chore(explorer): refactor chain icon and utilities
